requirements.txt
netifaces
psycopg2-binary
requests
pyyaml
